pyproject.toml
[project]
name = "bricky"
version = "0.1.0"
description = ""
authors = [
    {name = "kerem212012",email = "rockstarfoxykerem@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "ruff (>=0.14.3,<0.15.0)",
    "django-phonenumber-field[phonenumbers] (>=8.3.0,<9.0.0)",
    "django-debug-toolbar (>=6.1.0,<7.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "environs (>=14.5.0,<15.0.0)",
    "pytelegrambotapi (>=4.29.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["F401", "F841"]
